+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+* Use snake cased `history.to_href`
+
 ## 2.0.0-beta.1
 
 * Remove `automaticRedirects` option. `createServerHistory`, from `@hickory/in-memory`, makes `navigate` a no-op and should be used for server rendering.